Gophercises Ex 1 - Refactor After Checking Solution
I looked at the solution for part 1.

- use time.Timer
- create a unique channel for each answer
- the "listening" for the answer is done
in a separate go routine to the main routine
in the loop can react to the timer message
as soon as it's sent. This was something I was
stuck on when I created my original implementation.
I understood that I needed to react to the timer
immediately, but I reversed things a bit.
With this implementation, the problem loop will hold
until the SELECT hears from the timer or the answer
channel.