Exercise 2 - URL Shortener - Refactor
As far as I can see, there are two basic
ways to handle the different formats used
to configure the redirects.
You could create "handlers" for each type,
that do the work of parsing the configuration
and then creating a distinct map of redirects
just parsed...
or you could maintain a single map and just
append to this map after parsing each configuration
format.
In my first commit, I went with the first option.
After thinking about the root goal - redirects -
I think this last commit is much clearer.
We simply maintain a unique function for each format
that will parse the format and append the
redirects to the existing map. From there, there
is a single handler with the default mux used
as the fallback handler.
It's a subtle change, for sure, but a better one,
I think.